# triton-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Triton Backend",
    "description": "Backend for Triton",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/perf-analyzers": {
      "get": {
        "tags": [
          "perf-analyzers"
        ],
        "summary": "Get Perf Analyzer Status",
        "description": "Return the singleton Perf Analyzer installation state.",
        "operationId": "get_perf_analyzer_status_api_perf_analyzers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfAnalyzerStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "perf-analyzers"
        ],
        "summary": "Install Perf Analyzer",
        "description": "Install a Perf Analyzer SDK container on Kubernetes.",
        "operationId": "install_perf_analyzer_api_perf_analyzers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallPerfAnalyzerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfAnalyzerInstallResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "perf-analyzers"
        ],
        "summary": "Uninstall Perf Analyzer",
        "description": "Uninstall the singleton Perf Analyzer Kubernetes workload.",
        "operationId": "uninstall_perf_analyzer_api_perf_analyzers_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfAnalyzerDeleteResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/perf-analyzers/runs": {
      "post": {
        "tags": [
          "perf-analyzers"
        ],
        "summary": "Run Perf Analyzer",
        "description": "Execute Perf Analyzer for one Triton instance model.",
        "operationId": "run_perf_analyzer_api_perf_analyzers_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunPerfAnalyzerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfAnalyzerRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/perf-analyzers/runs/latest": {
      "get": {
        "tags": [
          "perf-analyzers"
        ],
        "summary": "Get Latest Perf Analyzer Run",
        "description": "Return the latest persisted Perf Analyzer result for one model target.",
        "operationId": "get_latest_perf_analyzer_run_api_perf_analyzers_runs_latest_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          },
          {
            "name": "model_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Name"
            }
          },
          {
            "name": "model_version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfAnalyzerLatestRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InstallPerfAnalyzerRequest": {
        "properties": {
          "installation_name": {
            "type": "string",
            "title": "Installation Name",
            "default": "perf-analyzer"
          },
          "image": {
            "type": "string",
            "title": "Image"
          },
          "dockerconfigjson": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dockerconfigjson"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "InstallPerfAnalyzerRequest",
        "description": "Request body for ``POST /api/perf-analyzers``."
      },
...
      "PerfAnalyzerDeleteResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "namespace": {
            "type": "string",
            "title": "Namespace"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message",
          "namespace"
        ],
        "title": "PerfAnalyzerDeleteResponse",
        "description": "Response after uninstalling the singleton Perf Analyzer workload."
      },
      "PerfAnalyzerInstallResponse": {
        "properties": {
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "deployment_name": {
            "type": "string",
            "title": "Deployment Name"
          },
          "image": {
            "type": "string",
            "title": "Image"
          },
          "applied_resources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Applied Resources"
          }
        },
        "type": "object",
        "required": [
          "namespace",
          "deployment_name",
          "image",
          "applied_resources"
        ],
        "title": "PerfAnalyzerInstallResponse",
        "description": "Response after Perf Analyzer Kubernetes resources are applied."
      },
      "PerfAnalyzerLatestRunResponse": {
        "properties": {
          "found": {
            "type": "boolean",
            "title": "Found"
          },
          "executed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executed At"
          },
          "batch_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Size"
          },
          "concurrency_range": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concurrency Range"
          },
          "measurement_request_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measurement Request Count"
          },
          "input_data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Data"
          },
          "command": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Command"
          },
          "output": {
            "type": "string",
            "title": "Output",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "found"
        ],
        "title": "PerfAnalyzerLatestRunResponse",
        "description": "Persisted latest Perf Analyzer result for one instance model target."
      },
      "PerfAnalyzerRunResponse": {
        "properties": {
          "command": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Command"
          },
          "output": {
            "type": "string",
            "title": "Output"
          }
        },
        "type": "object",
        "required": [
          "command",
          "output"
        ],
        "title": "PerfAnalyzerRunResponse",
        "description": "Output captured from a Perf Analyzer process executed in Kubernetes."
      },
      "PerfAnalyzerStatusResponse": {
        "properties": {
          "installed": {
            "type": "boolean",
            "title": "Installed"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "not_installed"
          },
          "ready": {
            "type": "boolean",
            "title": "Ready",
            "default": false
          },
          "status_message": {
            "type": "string",
            "title": "Status Message",
            "default": ""
          },
          "installation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PerfAnalyzerInstallResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "installed"
        ],
        "title": "PerfAnalyzerStatusResponse",
        "description": "Response describing the singleton Perf Analyzer installation state."
      },
      "RunPerfAnalyzerRequest": {
        "properties": {
          "instance_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Instance Id"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "batch_size": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Batch Size",
            "default": 1
          },
          "concurrency_range": {
            "type": "string",
            "title": "Concurrency Range",
            "default": "1"
          },
          "measurement_request_count": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Measurement Request Count",
            "default": 50
          },
          "input_data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Data"
          }
        },
        "type": "object",
        "required": [
          "instance_id",
          "model_name",
          "model_version"
        ],
        "title": "RunPerfAnalyzerRequest",
        "description": "Request body for a model-level Perf Analyzer run."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}